CNN-5 Added File Hirearchy
