+++ Ashhar/CNN_using_C/image.h
@@ -1,0 +1,6 @@
+struct Image{
+    // Declare a 2D array to store the image data.
+    float image_array[30][30];
+    int height;
+    int width;
+};
