+++ Ashhar/CNN_using_C/main.c
@@ -1,16 +1,12 @@
 #include "matrix.h"
+#include "image.h"
+
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
 #include <fcntl.h>
 
-
-#define STB_IMAGE_IMPLEMENTATION
-#include "stb/stb_image.h"
-#define STB_IMAGE_WRITE_IMPLEMENTATION
-#include "stb/stb_image_write.h"
-
-int img_index = 0;
+int img_index = 1;
 
 
 // extern void Matrix_Init(struct Matrix *x, int r, int c);
@@ -21,17 +17,12 @@ int img_index = 0;
 // extern struct Matrix multiply_matrices(struct Matrix *matrix1, struct Matrix *matrix2);
 // extern struct Matrix multiply_matrices(struct Matrix *matrix1, struct Matrix *matrix2);
 
-struct Image{
-    // Declare a 2D array to store the image data.
-    float image_array[30][30];
-    int height;
-    int width;
-};
+// function that will retrieve the images that are to be used for training the CNN
+extern void get_images(int per_num, struct Image* image);
 
 
 // functions for the convolutional layer
 void filter_init();
-void get_images(int);
 void convolution_forward(struct Image);
 
 
@@ -91,7 +82,7 @@ int main()
 
 
     // initialize the images for the training dataset
-    get_images(10);
+    get_images(10, image);
     
     // initialize the filter for the convolution layer
     filter_init();
@@ -128,77 +119,11 @@ int main()
     for(int i = 0; i < 3; i++){
         free(filter[i]);
     }
-    free(filter);]
+    free(filter);
 
     return 0;
 }
 
-
-
-
-
-
-
-// function to populate the "images" object of our Image structure (will generalize later on)
-void get_images(int per_num){
-    uint8_t img[28][28];
-    unsigned char* image_data;
-    
-    // Load the image data into the array.
-    image_data = stbi_load("/home/ashhar/Desktop/CNN_on_FPGA/Ashhar/CNN_using_C/mnist_png/training/0/6034.png", &image[0].width, &image[0].height, NULL, 0);
-
-    if (image_data == NULL) {
-        printf("\nError Opening The Image!\n");
-        // Handle error.
-        exit(-1);
-    }
-
-    memcpy(img, image_data, image[0].width * image[0].height);
-
-    // Pad the image with zeros
-    for (int i = 0; i < 28; i++) {
-        for (int j = 0; j < 28; j++) {
-            image[0].image_array[i][j] = 0;
-        }
-    }
-
-    // Pad the image with zeros
-    for (int i = 0; i < 28; i++) {
-        for (int j = 0; j < 28; j++) {
-            image[0].image_array[i+1][j+1] = img[i][j] / (float) 255;
-        }
-    }
-
-    image[0].width += 2;
-    image[0].height += 2;
-
-    // Load the image data into the array.
-    image_data = stbi_load("/home/ashhar/Desktop/CNN_on_FPGA/Ashhar/CNN_using_C/mnist_png/training/3/59957.png", &image[1].width, &image[1].height, NULL, 0);
-
-    if (image_data == NULL) {
-        printf("\nError Opening The Image!\n");
-        // Handle error.
-        exit(-1);
-    }
-
-    memcpy(img, image_data, image[1].width * image[1].height);
-
-    // Pad the image with zeros
-    for (int i = 0; i < 28; i++) {
-        for (int j = 0; j < 28; j++) {
-            image[1].image_array[i][j] = 0;
-        }
-    }
-
-    // Pad the image with zeros
-    for (int i = 0; i < 28; i++) {
-        for (int j = 0; j < 28; j++) {
-            image[1].image_array[i+1][j+1] = img[i][j] / (float)255;
-        }
-    }
-    image[1].width += 2;
-    image[1].height += 2;
-}
 
 
 // THESE ARE THE FUNCTION THAT ARE USED BY THE CONVOLUTIONAL LAYER:
