+++ Ashhar/CNN_using_C/makefile
@@ -1,2 +1,2 @@
 all:
-	gcc -o matrix matrix.c main.c -lpng -lm
+	gcc -o matrix matrix.c main.c image.c -lpng -lm
